Revert "Switched from doubles to decimals"
This reverts commit c364ae38ef330cd18b4713b10d49b3cda5c46c9f.

+++ Contracts/IExpression.cs
@@ -15,7 +15,7 @@ namespace Team1922.MVVM.Contracts
         /// executes this expression
         /// </summary>
         /// <returns>the result of the expression</returns>
-        decimal Evaluate();
+        double Evaluate();
         /// <summary>
         /// Gets the string representation of this expression
         /// </summary>
